More progress - At impasse
I've pretty much run out of ideas at this point. Going to meet with someone to see if I can find a solution. Not giving up yet, but I may be done for today.

Issue at this point is that the innerGridX only checks current grid. If you pass over to the tile to the left, you are no longer overlapping on the left, you are now overlapping on the right. If you cross back over, the square you entered then left immediately will stay marked as player.