Optimizations and fixes
Optimized the update loop to only update information if it has actually changed. Additionally, it only sends that information if it updates. On client side, the update for the player is skipped if there is no information changes. Also fixed respawning which could cause a server crash. Check changelog for updates.