Changed var/let declarations to const if fitting
JavaScript has const, and I felt it would be better if I made all non-changing variables const, which is about 90% of them honestly. Now there's the issue with Objects not being actually immutable, but I believe none of my objects actually need to be frozen.

Therefore, now almost all of my variables are const, meaning they will never be reassigned to some other value. This includes functions. It's a slight change, but I think makes a bit more sense when reading.

I already program with everything I used declared before it is used, so I have no issues there, the transition was just using Find and Replace and skipping over for loops and a variable here and there.